Remove obfuscator file
As part of cleaning up the overall code base and to avoid having one
extra file that has to be included into the FormHelper this commit moves
the obfuscator methods into the main Mead module.
